+++ lib/Wechat.php
@@ -10,16 +10,35 @@ namespace C\lib {
 
     class Wechat {
 
+        // --- 公众号登录 ---
         public static function login($url, $appid = NULL) {
 
             $appid = $appid ? $appid : WECHAT_APPID;
-            $url = 'http'.(self::isHttps() ? 's':'').'://'.HTTP_HOST.'/' . $url;
+            $lenUrl = substr($url, 0, 6);
+            if ($lenUrl != 'https:' && $lenUrl != 'http:/') {
+                $url = 'http' . (self::isHttps() ? 's' : '') . '://' . HTTP_HOST . '/' . $url;
+            }
             header('Location: //open.weixin.qq.com/connect/oauth2/authorize?appid='.$appid.'&redirect_uri=' . urlencode($url) . '&response_type=code&scope=snsapi_userinfo&state=STATE#wechat_redirect');
 
         }
 
-        public static function redirect($appid = NULL, $secret = NULL) {
+        // --- 公众号获取用户信息 ---
+        public static function getUserInfo($access_token, $openid) {
+            return json_decode(Net::get('https://api.weixin.qq.com/sns/userinfo?access_token='.$access_token.'&openid='.$openid));
+        }
 
+        // --- 小程序登录 ---
+        public static function loginMS($code, $appid = NULL, $secret = NULL) {
+            $appid = $appid ? $appid : WECHAT_APPID;
+            $secret = $secret ? $secret : WECHAT_SECRET;
+
+            $r = Net::get('https://api.weixin.qq.com/sns/jscode2session?appid='.$appid.'&secret='.$secret.'&js_code='.$code.'&grant_type=authorization_code');
+            $j = json_decode($r);
+            return $j;
+        }
+
+        // --- 登录回跳处理 ---
+        public static function redirect($appid = NULL, $secret = NULL) {
             $appid = $appid ? $appid : WECHAT_APPID;
             $secret = $secret ? $secret : WECHAT_SECRET;
             if(isset($_GET['code']) && $_GET['code'] != '') {
@@ -41,6 +60,7 @@ namespace C\lib {
 
         }
 
+        // --- 创建支付 ---
         public static function createPay($opt = []) {
 
             require LIB_PATH . 'WxpayAPI/lib/WxPay.Api.php';
@@ -68,8 +88,10 @@ namespace C\lib {
 
         }
 
+        // --- 支付回调 ---
         public static function payCallback() {
 
+            require LIB_PATH . 'WxpayAPI/lib/WxPay.Api.php';
             require LIB_PATH . 'WxpayAPI/lib/WxPay.Notify.php';
             require LIB_PATH . 'WxpayAPI/lib/WxPay.NotifyCallBack.php';
 
@@ -78,7 +100,30 @@ namespace C\lib {
 
         }
 
-        public static  function isHttps() {
+        // --- 获取服务器 Signature (7200秒刷新一次) ---
+        // --- ['onMenuShareTimeline', 'onMenuShareAppMessage'] ---
+        public static function getWXConfig($apiList, $tokenTicket, $appid = NULL, $secret = NULL) {
+            $appid = $appid ? $appid : WECHAT_APPID;
+            $secret = $secret ? $secret : WECHAT_SECRET;
+
+            if (!is_string($tokenTicket) || $tokenTicket == '') {
+                $r = Net::get('https://api.weixin.qq.com/cgi-bin/token?grant_type=client_credential&appid=' . $appid . '&secret=' . $secret);
+                $r = json_decode($r);
+                $q = Net::get('https://api.weixin.qq.com/cgi-bin/ticket/getticket?access_token=' . $r->access_token . '&type=jsapi');
+                $q = json_decode($q);
+                $ticket = $q->ticket;
+                $tokenTicket = $r->access_token . ',' . $ticket;
+            } else {
+                list($token, $ticket) = explode(',', $tokenTicket);
+            }
+            $noncestr = Text::random(16, ['L', 'U', 'N']);
+            $time = $_SERVER['REQUEST_TIME'];
+            $string = 'jsapi_ticket='.$ticket.'&noncestr='.$noncestr.'&timestamp='.$time.'&url=http' . ((self::isHttps() ? 's' : '') . '://') . $_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI'];
+            return [$tokenTicket, 'wx.config({debug:false,appId:"'.$appid.'",timestamp:"'.$time.'",nonceStr:"'.$noncestr.'",signature:"'.sha1($string).'",jsApiList:'.json_encode($apiList).'});'];
+        }
+
+        // --- 判断是否是 HTTPS ---
+        public static function isHttps() {
             if (isset($_SERVER['HTTPS'])) {
                 if ($_SERVER['HTTPS'] === 1) {  //Apache
                     return true;